更新

+++ src/main/java/com/bootdo/oa/service/impl/FjServiceImpl.java
@@ -4,7 +4,9 @@ import com.bootdo.common.utils.FileUtil;
 import com.bootdo.common.utils.UUIDUtils;
 import com.bootdo.oa.dao.FjDao;
 import com.bootdo.oa.domain.FjDO;
+import com.bootdo.oa.domain.JcxxDO;
 import com.bootdo.oa.service.FjService;
+import com.bootdo.oa.service.JcxxService;
 import io.netty.util.internal.StringUtil;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
@@ -23,6 +25,9 @@ public class FjServiceImpl implements FjService {
 
 	@Value("${bootdo.uploadPath}")
 	private String uploadPath;
+
+	@Autowired
+	private JcxxService jcxxService;
 	
 	@Override
 	public FjDO get(String id){
@@ -47,6 +52,13 @@ public class FjServiceImpl implements FjService {
 	
 	@Override
 	public int update(FjDO fj){
+		if (!StringUtil.isNullOrEmpty(fj.getStudyImg()) ||!StringUtil.isNullOrEmpty(fj.getDegreeImg()) ||
+		!StringUtil.isNullOrEmpty(fj.getCardImgF()) || !StringUtil.isNullOrEmpty(fj.getCardImgR())){
+			JcxxDO jcxx = jcxxService.get(fj.getJcxxId());
+			jcxx.setIsFj("1");//已上传附件
+			jcxxService.update(jcxx);
+		}
+
 		return fjDao.update(fj);
 	}
 	
@@ -65,20 +77,21 @@ public class FjServiceImpl implements FjService {
 		//删除数据库的同时把对面本地文件夹的图片删除
 		//先查出图片本地路径
 		FjDO fj = fjDao.getByJcxxId(jcxxId);
-		//删除图片
-		if (!StringUtil.isNullOrEmpty(fj.getStudyImg())){
-			FileUtil.deleteFile(fj.getStudyImg());
+		if (Objects.nonNull(fj)){
+			//删除图片
+			if (!StringUtil.isNullOrEmpty(fj.getStudyImg())){
+				FileUtil.deleteFile(fj.getStudyImg());
+			}
+			if (!StringUtil.isNullOrEmpty(fj.getDegreeImg())){
+				FileUtil.deleteFile(fj.getDegreeImg());
+			}
+			if (!StringUtil.isNullOrEmpty(fj.getCardImgF())){
+				FileUtil.deleteFile(fj.getCardImgF());
+			}
+			if (!StringUtil.isNullOrEmpty(fj.getCardImgR())){
+				FileUtil.deleteFile(fj.getCardImgR());
+			}
 		}
-		if (!StringUtil.isNullOrEmpty(fj.getDegreeImg())){
-			FileUtil.deleteFile(fj.getDegreeImg());
-		}
-		if (!StringUtil.isNullOrEmpty(fj.getCardImgF())){
-			FileUtil.deleteFile(fj.getCardImgF());
-		}
-		if (!StringUtil.isNullOrEmpty(fj.getCardImgR())){
-			FileUtil.deleteFile(fj.getCardImgR());
-		}
-
 		return fjDao.deleteByJcxxId(jcxxId);
 	}
 
@@ -88,16 +101,30 @@ public class FjServiceImpl implements FjService {
 		for (MultipartFile file : files) {
 			if (Objects.nonNull(file)) {
 				try {
-					if (file.getName().equals("毕业证")) {
+					String originalFilename = file.getOriginalFilename();
+					String fileName = originalFilename.replaceAll("\\..*$","");
+					if (fileName.equals("毕业证")) {
+						if (!StringUtil.isNullOrEmpty(fj.getStudyImg())){
+							FileUtil.deleteFile(fj.getStudyImg());
+						}
 						String studyImg = FileUtil.uploadFile(file.getBytes(), uploadPath, UUIDUtils.randomUUID());
 						fj.setStudyImg(studyImg);
-					} else if (file.getName().equals("学位证")) {
+					} else if (fileName.equals("学位证")) {
+						if (!StringUtil.isNullOrEmpty(fj.getDegreeImg())){
+							FileUtil.deleteFile(fj.getDegreeImg());
+						}
 						String degreeImg = FileUtil.uploadFile(file.getBytes(), uploadPath, UUIDUtils.randomUUID());
 						fj.setDegreeImg(degreeImg);
-					} else if (file.getName().equals("身份证前面")) {
+					} else if (fileName.equals("身份证前面")) {
+						if (!StringUtil.isNullOrEmpty(fj.getCardImgF())){
+							FileUtil.deleteFile(fj.getCardImgF());
+						}
 						String cardImgF = FileUtil.uploadFile(file.getBytes(), uploadPath, UUIDUtils.randomUUID());
 						fj.setCardImgF(cardImgF);
-					} else if (file.getName().equals("身份证反面")) {
+					} else if (fileName.equals("身份证反面")) {
+						if (!StringUtil.isNullOrEmpty(fj.getCardImgR())){
+							FileUtil.deleteFile(fj.getCardImgR());
+						}
 						String cardImgR = FileUtil.uploadFile(file.getBytes(), uploadPath, UUIDUtils.randomUUID());
 						fj.setCardImgR(cardImgR);
 					}
@@ -107,6 +134,11 @@ public class FjServiceImpl implements FjService {
 			}
 		}
 		update(fj);
+	}
+
+	@Override
+	public FjDO getByJcxxId(String jcxxId) {
+		return fjDao.getByJcxxId(jcxxId);
 	}
 
 }
